LayoutTests/fast/events/popup-blocked-from-fake-button-click.html

<html> 
    <head>
        <script src="../js/resources/js-test-pre.js"></script>
    </head>
    <body>
        <button id="test" onclick="popup()" style="display:none"></button>
        <div id="console"></div>
        <script> 
            var win;
            function popup() {
                win = window.open("about:blank", "blank");
                shouldBeUndefined("win");
            }

            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.setCanOpenWindows();
                layoutTestController.setPopupBlockingEnabled(true);
                layoutTestController.setCloseRemainingWindowsWhenComplete(true);
                layoutTestController.waitUntilDone();
            }

            document.getElementById("test").click();

            if (window.layoutTestController)
                layoutTestController.notifyDone();
        </script> 
    </body> 
</html>
